Fix in memory transact race condition in Clojure
This patch prevents a race condition that can occur when multiple
transactions are executed concurrently.

Previously the in memory implementation of `transact-update*` would
get `db-before` and the next `tx-id` via `db*` and `next-tx*`
respectively. Both of these functions read data out of the
connection state, an atom. When called concurrently, this can be
problematic because both values are used to determine what is
written to the state and may be stale by that time causing successive
transactions to clobber one another.

To solve this problem we use `swap!` to ensure the both the current
state of the database and history are read at the time of writing.